LayoutTests/fast/dom/shadow/cppevent-input-in-shadow-expected.txt

host: copy is fired.
dark: copy is fired.
host: cut is fired.
dark: cut is fired.
host: paste is fired.
dark: paste is fired.
PASS successfullyParsed is true

TEST COMPLETE
